fix(signedness): Change variables signedness to match the operation
The flag -wextra forces comparisons and other operations to have elements
with the same signedness. In the commit, we change some variables types to
match the operations they are used, or cast them to the correct type.

+++ src/arch/riscv/irqc/aia/vaplic.c
@@ -260,7 +260,7 @@ static void vaplic_update_hart(struct vcpu* vcpu, int16_t vhart_index)
         for (size_t i = 0; i < vaplic->idc_num; i++) {
             vaplic_update_hart_line(vcpu, (vcpuid_t)i);
         }
-    } else if (vhart_index < vaplic->idc_num) {
+    } else if ((uint16_t)vhart_index < vaplic->idc_num) {
         vaplic_update_hart_line(vcpu, (vcpuid_t)vhart_index);
     }
 }
